myapp/requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.8.0     
django-otp==1.6.1
django-phone-field==1.8.1      
django-phonenumber-field==8.1.0
djangorestframework==3.16.1    
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
phonenumbers==9.0.13
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2